Make search bar a part of header
