+++ packages/gatsby/src/components/layout.js
@@ -13,11 +13,11 @@ import React     from 'react';
 
 import Header    from './header';
 
-const Layout = ({ children }) => <>
+const Layout = ({ header, children }) => <>
   <Helmet>
     <meta name={`viewport`} content={`width=device-width, initial-scale=1`} />
   </Helmet>
-  <Header/>
+  {header ? header : <Header/>}
   <main>
     {children}
   </main>
